код разбит на модули. Версия 1
