+++ js/functions.mjs
@@ -42,5 +42,3 @@ const extractingNumbers = (string = '') => {
 // console.log(extractingNumbers(2023));
 // console.log(extractingNumbers(-1));
 // console.log(extractingNumbers(1.5));
-
-
