+++ js/main.mjs
@@ -1,0 +1,8 @@
+import { createPhotoDescriptions } from './photo-description.mjs';
+
+const NUMBER_OF_REQUIRED_OBJECTS = 25;
+
+const photoDescriptions = createPhotoDescriptions(NUMBER_OF_REQUIRED_OBJECTS);
+
+console.log(photoDescriptions);
+
