+++ js/photo-data.mjs
@@ -1,0 +1,38 @@
+const DESCRIPTIONS = [
+  'Хорошая красивая фотография',
+  'Чувствую себя Анри Картье-Брессоном',
+  'Живите с этим',
+  'Павук',
+  'Вредный гном',
+  'Очень вредный гном',
+  'Просто гном',
+  'Солнышко проснулось - новый день на дворе',
+  'Это я в Дубае, я ща отдыхаю',
+  'Муж купил, мне муж купил'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Иов',
+  'Агафопод',
+  'Феодул',
+  'Алексий',
+  'Николай',
+  'Пуплий',
+  'Феона',
+  'Симеон',
+  'Форвин',
+  'Марк',
+  'Платон',
+  'Феодора'
+];
+
+export { DESCRIPTIONS, MESSAGES, NAMES };
